#6 IDをLongに変更

+++ domain/src/main/kotlin/domain/entity/user/User.kt
@@ -3,7 +3,7 @@ package domain.entity.user
 import domain.entity.IEntity
 
 data class User(
-    override val id: Int = -1,
+    override val id: Long = -1,
     val firstName: String,
     val lastName: String,
     val email: Email
